backend/requirements.txt
fastapi==0.115.9
uvicorn>=0.34.3
pydantic>=2.11.5
sqlalchemy>=2.0.41
psycopg2-binary>=2.9.9
python-multipart>=0.0.20
alembic>=1.16.1
pydantic-settings>=2.9.1
httpx>=0.28.1
pymupdf>=1.26.0
numpy<2.0.0
sentence-transformers>=4.1.0
chromadb>=1.0.12
pytest>=8.4.0